+++ src/main/java/it/montano/multipersistencebackend/product/service/ProductPostgresService.java
@@ -49,6 +49,9 @@ public class ProductPostgresService implements ProductService {
   @Transactional
   @Override
   public void deleteProduct(@NonNull UUID productId) {
+    if (!repo.existsById(productId)) {
+      throw new ResourceNotFoundException("Product not found with id " + productId);
+    }
     repo.deleteById(productId);
   }
 
@@ -64,7 +67,7 @@ public class ProductPostgresService implements ProductService {
   public @NonNull ProductResponse getProductById(@NonNull UUID productId) {
     return repo.findById(productId)
         .map(mapper::toResponse)
-        .orElseThrow(() -> new ResourceNotFoundException(productId.toString()));
+        .orElseThrow(() -> new ResourceNotFoundException("Product not found with id " + productId));
   }
 
   /**
@@ -92,7 +95,8 @@ public class ProductPostgresService implements ProductService {
       @NonNull UUID productId, @NonNull ProductRequest productRequest) {
     ProductEntity entity =
         repo.findById(productId)
-            .orElseThrow(() -> new ResourceNotFoundException(productId.toString()));
+            .orElseThrow(
+                () -> new ResourceNotFoundException("Product not found with id " + productId));
     mapper.updateEntity(productRequest, entity);
     return mapper.toResponse(repo.save(entity));
   }
